Tidy up merge conflicts
