+++ yoti_client_test.go
@@ -12,6 +12,7 @@ import (
 	"io/ioutil"
 	"net/http"
 	"net/url"
+	"os"
 	"strings"
 	"testing"
 	"time"
@@ -201,7 +202,7 @@ func TestYotiClient_SharingFailure_ReturnsFailure(t *testing.T) {
 func TestYotiClient_TokenDecodedSuccessfully(t *testing.T) {
 	key, _ := ioutil.ReadFile("test-key.pem")
 
-	expectedAbsoluteURL := "/api/v1/profile/" + encryptedToken
+	expectedAbsoluteURL := "/api/v1/profile/" + token
 
 	client := Client{
 		HTTPClient: &mockHTTPClient{
@@ -578,4 +579,45 @@ func TestClient_OverrideAPIURL_ShouldSetAPIURL(t *testing.T) {
 	expectedURL := "expectedurl.com"
 	client.OverrideAPIURL(expectedURL)
 	assert.Equal(t, client.getAPIURL(), expectedURL)
+}
+
+func TestYotiClient_GetAPIURLUsesOverriddenBaseUrlOverEnvVariable(t *testing.T) {
+	client := Client{}
+	client.OverrideAPIURL("overridenBaseUrl")
+
+	os.Setenv("YOTI_API_URL", "envBaseUrl")
+
+	result := client.getAPIURL()
+
+	assert.Equal(t, "overridenBaseUrl", result)
+}
+
+func TestYotiClient_GetAPIURLUsesEnvVariable(t *testing.T) {
+	client := Client{}
+
+	os.Setenv("YOTI_API_URL", "envBaseUrl")
+
+	result := client.getAPIURL()
+
+	assert.Equal(t, "envBaseUrl", result)
+}
+
+func TestYotiClient_GetAPIURLUsesDefaultUrlAsFallbackWithEmptyEnvValue(t *testing.T) {
+	client := Client{}
+
+	os.Setenv("YOTI_API_URL", "")
+
+	result := client.getAPIURL()
+
+	assert.Equal(t, "https://api.yoti.com/api/v1", result)
+}
+
+func TestYotiClient_GetAPIURLUsesDefaultUrlAsFallbackWithNoEnvValue(t *testing.T) {
+	client := Client{}
+
+	os.Unsetenv("YOTI_API_URL")
+
+	result := client.getAPIURL()
+
+	assert.Equal(t, "https://api.yoti.com/api/v1", result)
 }
