--- yoticlient_test.go
@@ -1,49 +1,0 @@
-package yoti
-
-import (
-	"os"
-	"testing"
-
-	"gotest.tools/v3/assert"
-)
-
-func TestYotiClient_GetAPIURLUsesOverriddenBaseUrlOverEnvVariable(t *testing.T) {
-	client := Client{}
-	client.OverrideAPIURL("overridenBaseUrl")
-
-	os.Setenv("YOTI_API_URL", "envBaseUrl")
-
-	result := client.getAPIURL()
-
-	assert.Equal(t, "overridenBaseUrl", result)
-}
-
-func TestYotiClient_GetAPIURLUsesEnvVariable(t *testing.T) {
-	client := Client{}
-
-	os.Setenv("YOTI_API_URL", "envBaseUrl")
-
-	result := client.getAPIURL()
-
-	assert.Equal(t, "envBaseUrl", result)
-}
-
-func TestYotiClient_GetAPIURLUsesDefaultUrlAsFallbackWithEmptyEnvValue(t *testing.T) {
-	client := Client{}
-
-	os.Setenv("YOTI_API_URL", "")
-
-	result := client.getAPIURL()
-
-	assert.Equal(t, "https://api.yoti.com/api/v1", result)
-}
-
-func TestYotiClient_GetAPIURLUsesDefaultUrlAsFallbackWithNoEnvValue(t *testing.T) {
-	client := Client{}
-
-	os.Unsetenv("YOTI_API_URL")
-
-	result := client.getAPIURL()
-
-	assert.Equal(t, "https://api.yoti.com/api/v1", result)
-}
